+++ libs/components/router/testing/src/href/fixtures/href-harness-test.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { SkyHrefModule } from '@skyux/router';
+
+import { HrefHarnessTestComponent } from './href-harness-test.component';
+
+@NgModule({
+  declarations: [HrefHarnessTestComponent],
+  exports: [HrefHarnessTestComponent],
+  imports: [SkyHrefModule],
+})
+export class HrefHarnessTestModule {}